optimizing code

+++ Projects/Eraswap/contracts/Staking.sol
@@ -11,7 +11,7 @@ contract Staking {
 
     // Event to watch staking creations
     event stakeCreation(
-    uint256 orderid,
+    uint64 orderid,
     address indexed ownerAddress,
     uint256 value
     );
@@ -19,12 +19,12 @@ contract Staking {
 
     // Event to watch loans repayed taken
     event loanTaken(
-    uint256 orderid
+    uint64 orderid
     );
 
     // Event to watch wind up of contracts
     event windupContract(
-    uint256 orderid
+    uint64 orderid
     );
 
     IERC20   tokenContract;  // Defining conract address so as to interact with EraswapToken
